refactor: extract GUI subsystem into `solvent` crate
The kernel's GUI module (`gui.rs`) has been stripped down to a thin bridge that provides framebuffer access and GPU present/flush, while all rendering and compositing logic is now handled by the new `solvent` crate. Dependencies on `lattice`, `resonance`, and `chronoline` are removed from the kernel and replaced with a single `solvent` dependency. This improves modularity by isolating GUI functionality into its own crate.
