+++ fullerene-kernel/Cargo.toml
@@ -21,9 +21,7 @@ spin = { version = "0.10.0", default-features = false, features = [
 petroleum = { path = "../petroleum" }
 nitrogen = { path = "../nitrogen" }
 nozzle = { path = "../nozzle" }
-lattice = { path = "../lattice" }
-resonance = { path = "../resonance" }
-chronoline = { path = "../chronoline" }
+solvent = { path = "../solvent" }
 embedded-graphics = { version = "0.8", default-features = false }
 memoffset = { version = "0.9", default-features = false }
 linked_list_allocator = { version = "0.10.6", default-features = false, features = [
